hebpsh/pyproject.toml
[tool.poetry]
name = "hebpsh"
version = "0.1.0"
description = "URL Shortener Service"
authors = ["Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.124.2"
sqlalchemy = "^2.0.45"
psycopg2-binary = "^2.9.11"
alembic = "^1.17.2"
python-dotenv = "^1.2.1"
uvicorn = "^0.38.0"
pydantic = "^2.12.5"
click = "^8.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
